stupid animation

+++ Tempest/src/Graphics/PBR/MeshPBR.h
@@ -24,7 +24,7 @@ namespace Tempest
         glm::vec3 Position;
         glm::vec3 Normal;
         glm::vec2 TexCoords;
-        glm::ivec4 BoneIds;
+        glm::vec4 BoneIds;
         glm::vec4 Weights;
     };
